[Player] Limit the playing field to the camera's view
The player cannot exit the playing field i.e. it cannot exit the main
camera's view.

The player's movement and collision detection with the playing field
borders has been achieved via Unity's built-in physics engine. The player
movement script has been altered to use said physics system.

A singleton game manager class was to be used to implement the changes.
However, it is not needed as of this time. It has been left for future
development.