[Level/Demo] Add Vi motion projectiles
This type of projectile applies a movement command similar to that of
Vi's movement. When hit by one of these projectiles, the player is
displaced in one of 4 directions. The displacement effect is multiplied
by a count (in the range [1;3]).

Also, sprites and fonts have been moved to a folder called 'Resources/'.
This is required in order to use the 'Resources.Load()' method in Unity.