[Refactoring] Reimplement movement field's follower behaviour
The 'GameObjectFollower' component has been redesigned into an abstract
class. It has a few abstract methods which should be implemented by
subclasses in order to achieve the follower behaviour.

The follower class is implemented by the new 'SmoothDampGameObjectFollower'
subclass which achieves the follower functionality by smoothly slithering
towards its target. It also defines a stop zone, which stops the
follower when entered.